revise in response to comments
