+++ manubot/process/tests/test_jinja.py
@@ -1,5 +1,5 @@
 def test_mb_random_order():
-    from manubot.filters import mb_random
+    from manubot.process.jinja import mb_random
     start = [0, 1, 2, 3, 4, 5, 6, 7, 8, 9]
     shuffled = mb_random(start, seed=5)
 
@@ -8,7 +8,7 @@ def test_mb_random_order():
 
 
 def test_mb_random_same_seed():
-    from manubot.filters import mb_random
+    from manubot.process.jinja import mb_random
     start = [0, 1, 2, 3, 4, 5, 6, 7, 8, 9]
     shuffled1 = mb_random(start, seed=5)
     shuffled2 = mb_random(start, seed=5)
@@ -17,7 +17,7 @@ def test_mb_random_same_seed():
 
 
 def test_mb_random_start_unchanged():
-    from manubot.filters import mb_random
+    from manubot.process.jinja import mb_random
     start = [0, 1, 2, 3, 4, 5, 6, 7, 8, 9]
     mb_random(start, seed=5)
 
@@ -25,9 +25,22 @@ def test_mb_random_start_unchanged():
 
 
 def test_mb_random_diff_seeds():
-    from manubot.filters import mb_random
+    from manubot.process.jinja import mb_random
     start = [0, 1, 2, 3, 4, 5, 6, 7, 8, 9]
     shuffled1 = mb_random(start, seed=5)
     shuffled2 = mb_random(start, seed=6)
+
+    assert shuffled1 != shuffled2
+
+
+def test_mb_random_no_seeds():
+    """
+    This will fail incorrectly if, by chance, the two shuffled lists end up
+    having the same order. With 10 entries it shouldn't happen too often.
+    """
+    from manubot.process.jinja import mb_random
+    start = [0, 1, 2, 3, 4, 5, 6, 7, 8, 9]
+    shuffled1 = mb_random(start)
+    shuffled2 = mb_random(start)
 
     assert shuffled1 != shuffled2
